% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scholar_id}
\alias{get_scholar_id}
\title{Search for Google Scholar ID by name and affiliation}
\usage{
get_scholar_id(first_name, last_name, affiliation = NA)
}
\arguments{
\item{first_name}{Researcher first name.}

\item{last_name}{Researcher last name.}

\item{affiliation}{Researcher affiliation.}
}
\value{
Google Scholar ID as a character string.
}
\description{
Search for Google Scholar ID by name and affiliation
}
\examples{
 get_scholar_id(first_name = "gordon", last_name = "getzinger")


}
