% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{n_followers}
\alias{n_followers}
\title{Get the number of users for a given twitter user}
\usage{
n_followers(twitter_handle)
}
\arguments{
\item{twitter_handle}{A twitter handle}
}
\value{
The number of users or NA when no user is found.
}
\description{
Get the number of users for a given twitter user
}
\examples{
  n_followers(twitter_handle = "gjgetzinger")

}
